app/requirements.txt
# FastAPI y dependencias
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Procesamiento de audio
faster-whisper==0.9.0
pydub==0.25.1
ffmpeg-python==0.2.0

# Procesamiento de lenguaje natural
transformers==4.35.2
torch==2.1.1
sentencepiece==0.1.99
protobuf==4.25.1

# Caché y optimización
redis==5.0.1
python-memcached==1.59

# Monitoreo y logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Seguridad
cryptography==41.0.5
python-dotenv==1.0.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
aiofiles==23.2.1
